pyproject.toml
[project]
name = "sse-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.7.1",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
[tool.setuptools]
py-modules = ["terminal_server_sse"]
